to double

+++ jzy3d-api/src/api/org/jzy3d/chart/Chart.java
@@ -378,7 +378,7 @@ public class Chart {
         return new Scale(view.getBounds().getZmin(), view.getBounds().getZmax());
     }
 
-    public float flip(double y) {
+    public double flip(double y) {
         return canvas.getRendererHeight() - y;
     }
